Accurate timer and clock
I have finished the timer and clock part of the project. However, the code is quite bulky. Sometime in the future, someone can make a pull request to reduce the amount of code for this commit then enter it in manually (do not merge) and commit to the main branch.